add systick

+++ include/rcc.hpp
@@ -7,5 +7,5 @@ namespace rcc
     void RCC_init(void);
 
     void SYSTICK_init(void);
-    uint32_t getSystick();
+    uint64_t getSystick();
 }
